form: Add a perfdata model
This is such a special case, it warrants a class on its own. The
validation part needs to properly identify its partner field's host
or service, and that the wanted perf data source actually exists.

+++ test/unit_test/tests/form_orm_Test.php
@@ -17,19 +17,9 @@ class Form_ORM_Test extends PHPUnit_Framework_TestCase {
 				)
 			)
 		)));
+	}
 
-		$tables = array(
-			'ORMDriverLS default' => array(
-				'contacts' => array(
-					array( 'name' => 'Someone' ),
-					array( 'name' => 'Tomtenisse' )
-				)
-			)
-		);
-
-		if($this->mock_data_path !== false) {
-			unlink($this->mock_data_path);
-		}
+	private function mock_orm_tables(array $tables) {
 		$this->mock_data_path = __DIR__ . '/' . $this->getName(false) . '.json';
 		file_put_contents($this->mock_data_path, json_encode($tables));
 		foreach($tables as $driver => $tables) {
@@ -41,8 +31,10 @@ class Form_ORM_Test extends PHPUnit_Framework_TestCase {
 	}
 
 	protected function tearDown() {
-		unlink($this->mock_data_path);
-		$this->mock_data_path = false;
+		if($this->mock_data_path !== false) {
+			unlink($this->mock_data_path);
+			$this->mock_data_path = false;
+		}
 		op5objstore::instance()->mock_clear();
 	}
 
@@ -60,6 +52,15 @@ class Form_ORM_Test extends PHPUnit_Framework_TestCase {
 	 * @expectedExceptionMessage da_contact does not point at a valid object
 	 */
 	public function test_process_fail() {
+		$tables = array(
+			'ORMDriverLS default' => array(
+				'contacts' => array(
+					array( 'name' => 'Someone' ),
+					array( 'name' => 'Tomtenisse' )
+				)
+			)
+		);
+		$this->mock_orm_tables($tables);
 		$form = $this->get_form();
 
 		$result = $form->process_data(array(
@@ -73,6 +74,15 @@ class Form_ORM_Test extends PHPUnit_Framework_TestCase {
 	 * and that it is put back as default value for the next form
 	 */
 	public function test_process() {
+		$tables = array(
+			'ORMDriverLS default' => array(
+				'contacts' => array(
+					array( 'name' => 'Someone' ),
+					array( 'name' => 'Tomtenisse' )
+				)
+			)
+		);
+		$this->mock_orm_tables($tables);
 		$form_for_processing = $this->get_form();
 		$form_with_defaults = $this->get_form();
 
@@ -93,5 +103,72 @@ class Form_ORM_Test extends PHPUnit_Framework_TestCase {
 		$content = $form_with_defaults->get_view()->render(false);
 		$this->assertContains('value="Someone"', $content);
 		$this->assertNotContains('value="Tomtenisse"', $content);
+	}
+
+	public function test_perfdata_option_successful_validation_depending_on_orm_object() {
+		$tables = array(
+			'ORMDriverLS default' => array(
+				'hosts' => array(
+					array(
+						'name' => 'Sueridus',
+						'perf_data_raw' => 'pkt=1;0;0;0;5 rta=0.007;2000.000;2000.000;; pl=0%;95;100;;'
+					)
+				)
+			)
+		);
+		$this->mock_orm_tables($tables);
+
+		$form = new Form_Model(
+			'pump action',
+			array(
+				new Form_Field_ORMObject_Model('host', 'Which host do you want to see perfdata for?', array('hosts')),
+				new Form_Field_Perfdata_Model('host_perfdata', 'Host perfdata source', 'host')
+			)
+		);
+
+		$result = $form->process_data(array(
+			'host' => array(
+				'value' => 'Sueridus',
+				'table' => 'hosts',
+			),
+			'host_perfdata' => 'pkt'
+		));
+
+		$this->assertSame('pkt', $result['host_perfdata']);
+	}
+
+	/**
+	 * @expectedException FormException
+	 * @expectedExceptionMessage The performance data source 'Munny Saelee' is not found on the given object
+	 */
+	public function test_perfdata_option_failing_validation_depending_on_orm_object() {
+		$tables = array(
+			'ORMDriverLS default' => array(
+				'hosts' => array(
+					array(
+						'name' => 'Sueridus',
+						'perf_data_raw' => 'pkt=1;0;0;0;5 rta=0.007;2000.000;2000.000;; pl=0%;95;100;;'
+					)
+				)
+			)
+		);
+		$this->mock_orm_tables($tables);
+
+		$form = new Form_Model(
+			'pump action',
+			array(
+				new Form_Field_ORMObject_Model('host', 'Which host do you want to see perfdata for?', array('hosts')),
+				new Form_Field_Perfdata_Model('host_perfdata', 'Host perfdata source', 'host')
+			)
+		);
+
+		$form->process_data(array(
+			'host' => array(
+				'value' => 'Sueridus',
+				'table' => 'hosts',
+			),
+			'host_perfdata' => 'Munny Saelee'
+		));
+		$this->assertTrue(false, "process_data() should have thrown an exception");
 	}
 }
